Reorganize SamplesHeatmapView.
Created SamplesHeatmapHeader and SamplesHeatmapControls component.
Reordered remaining functions in SamplesHeatmapView in a more intuitive fashion.

Fixed issue where background wasn't being sent to back-end when fetching heatmap data.
