+++ app/assets/src/components/views/compare/SamplesHeatmapView/index.js
@@ -1,0 +1,3 @@
+import SamplesHeatmapView from "./SamplesHeatmapView";
+
+export default SamplesHeatmapView;
