+++ app/assets/src/components/views/compare/SamplesHeatmapView/samples_heatmap_view.scss
